change access modifier and move WaitProvider to utils
